requirements.txt
aiocontextvars==0.2.2
asgiref==3.8.1
backoff==2.2.1
boto3==1.34.90
botocore==1.34.90
Brotli==1.1.0
CacheControl==0.14.0
cachetools==5.3.3
python-qpid-proton==0.39.0
centribal-platform-utils==0.5.0
centribal-platform-logger==0.3.1
certifi==2024.2.2
chardet==5.2.0
contextvars==2.4
Deprecated==1.2.14
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
ffmpeg-python==0.2.0
firebase-admin==6.5.0
future==1.0.0
gevent==24.2.1
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
grpcio-status==1.62.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
immutables==0.20
jmespath==1.0.1
msgpack==1.0.8
mysql-connector-python==8.4.0
mysqlclient==2.2.4
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-distro==0.45b0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-botocore==0.45b0
opentelemetry-instrumentation-dbapi==0.45b0
opentelemetry-instrumentation-django==0.45b0
opentelemetry-instrumentation-grpc==0.45b0
opentelemetry-instrumentation-logging==0.45b0
opentelemetry-instrumentation-mysql==0.45b0
opentelemetry-instrumentation-pika==0.45b0
opentelemetry-instrumentation-requests==0.45b0
opentelemetry-instrumentation-urllib3==0.45b0
opentelemetry-instrumentation-wsgi==0.45b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
packaging==24.0
pandas==2.2.2
phonenumbers==8.13.35
pika==1.3.2
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1-modules==0.4.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0
pytz==2024.1
py-spy==0.3.14
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
websockets==12.0
wrapt==1.16.0
XlsxWriter==3.2.0
zope.event==5.0
zope.interface==6.3
